Split SettingsPreferences
